refactor: Classic Game Logic
-Refactor game over menu logic into it's own UI component.

-Refactor placeholder stats used for html screen capture to be it's own independant component.
